// src/api/types/productType.ts
export interface productType {
  sold: number
  images: string[]
  subcategory: Subcategory[]
  ratingsQuantity: number
  _id: string
  title: string
  slug: string
  description: string
  quantity: number
  price: number
  imageCover: string
  category: Category
  brand: Brand
  ratingsAverage: number
  createdAt: string
  updatedAt: string
  id: string
  priceAfterDiscount?: number
  availableColors?: any[]
}

export interface Subcategory {
  _id: string
  name: string
  slug: string
  category: string
}

export interface Category {
  _id: string
  name: string
  slug: string
  image: string
  createdAt?:string
  updatedAt?:string
}

export interface Brand {
  _id: string
  name: string
  slug: string
  image: string
}
export interface Review {
  _id: string;
  review: string;
  rating: number;
  user: {
    _id: string;
    name: string;
  };
  createdAt: string;
}